[RIKU-8] refactor: MainTabLayout 내부 의존되는 자식 페이지 수정 및 pages 폴더로 페이지 tsx 파일 이동
