+++ src/pages/NewMain.tsx
@@ -119,7 +119,7 @@ const NewMain: React.FC = () => {
   useEffect(() => {
     const fetchMain = async () => {
       try {
-        const token = JSON.parse(localStorage.getItem('accessToken') || 'null');
+        const token = localStorage.getItem('accessToken') || 'null';
         const response = await customAxios.get(`/run`, {
           headers: { Authorization: `${token}` },
         });
